imx8m: dram: Add an empty entry for OPTEE at MMU table
Fix boot hang by adding DRAM2 entry to MMU table when using OP-TEE

Trucrux boards dynamically retrieve DRAM size from the EEPROM, which means the PHYS_SDRAM_2_SIZE macro is not defined.
This causes an issue when the board has more than 3GB of RAM, as the MMU entry reserved for OP-TEE is used for DRAM2 instead, leading to a boot hang.